move variable validation hook to hooks folder
